fix(icon-button): missing property
